requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Machine Learning
scikit-learn==1.5.2

# AI / LLM
google-genai==1.9.0

# Report Generation
jinja2==3.1.4
reportlab==4.2.5
python-pptx==1.0.2
markdown==3.7
beautifulsoup4==4.12.3

# File Handling
python-multipart==0.0.12

# Environment Variables
python-dotenv==1.0.1
